chore: [#47] 불필요하게 정의해놓은 CalendarView 제거 및 PeriodSelectVC에 해당 코드들 추가
